Drone Polish
Added code for decals when the drone crashes into the ground. Dont have the asset yet so it wont do anything currently. Added better enemy detection to the drone so it only evaluates x/z positions and ignores y values. Commented right now. Added logic for the drone to go higher when its near the core.